// src/MyProfile/Interfaces/IProfileInfo.ts
export interface IProfileInfo {
    authenticationMethod: string;
    creationDate: string;
    creator: string;
    fullName: string;
    id: string;
    isActive: boolean;
    IsDisabled: boolean;
    isSystemAdministrator: boolean;
    language: string;
    lastLoginDate: Date;
    lastWriteDate: Date;
    mbox: string;
    origin: string;
    role: string;
    username: string;
    department: string;
    manager: string;
    office: string
    phone: string
    skype: string
    title: string
};